y dtype=torch.long

+++ dataset.py
@@ -100,7 +100,7 @@ class MelSpecEncoded:
         root_dir = './data'
         if indices is not None:
             self.data = torch.tensor([]).to(device)
-            self.y = torch.tensor([]).to(device)
+            self.y = torch.tensor([], dtype=torch.long).to(device)
             for prefix in prefixes:
                 with open(os.path.join(root_dir, f'{prefix}_features.pt'), 'rb') as f:
                     self.data = torch.cat((self.data, torch.load(f, map_location=device)))
